requirements.txt
torch
torchvision
ultralytics
streamlit
opencv-python
numpy
Pillow
